+++ vector/src/main/java/im/vector/fragments/VectorSettingsPreferencesFragment.kt
@@ -2526,7 +2526,7 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
                             }
                             else -> {
                                 exportButton.isEnabled = false
-                                passPhrase2EditText.error = getString(R.string.encryption_export_passphrase_dont_match)
+                                passPhrase2EditText.error = getString(R.string.passphrase_passphrase_does_not_match)
                             }
                         }
                     }
@@ -2616,8 +2616,6 @@ class VectorSettingsPreferencesFragment : PreferenceFragmentCompat(), SharedPref
 
                 }
             })
-
-            importButton.isEnabled = false
 
             val importDialog = builder.show()
             val appContext = thisActivity.applicationContext